% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lookup.R
\name{build_lookup_from_fastas}
\alias{build_lookup_from_fastas}
\title{Create these lookup tables within in a given directory that contains FASTA files:
\itemize{
\item lookup.csv
\item lookup_from_tenx.csv
\item lookup_from_adaptive.csv
}}
\usage{
build_lookup_from_fastas(data_dir)
}
\arguments{
\item{data_dir}{A string, the directory containing FASTA files.}
}
\value{
Nothing.
}
\description{
Create these lookup tables within in a given directory that contains FASTA files:
\itemize{
\item lookup.csv
\item lookup_from_tenx.csv
\item lookup_from_adaptive.csv
}
}
