% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lookup.R
\name{extract_imgt_genes}
\alias{extract_imgt_genes}
\title{Extract gene names from all reference FASTA files in a folder.}
\usage{
extract_imgt_genes(data_dir)
}
\arguments{
\item{data_dir}{A string, the path to directory containing FASTA files.}
}
\value{
A dataframe of gene names.
}
\description{
Extract gene names from all reference FASTA files in a folder.
}
