% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_example_path}
\alias{get_example_path}
\title{Get full path to an example file or directory}
\usage{
get_example_path(file_name)
}
\arguments{
\item{file_name}{A string, the name of the example file or directory.}
}
\value{
A string, the path to example file or directory.
}
\description{
\code{get_example_path()} takes a file or folder name that is expected to be
located under the \code{TCRconvertR} \code{examples} directory and gets the full path
to that item.
}
\examples{
get_example_path("tenx.csv")
}
