% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lookup.R
\name{parse_imgt_fasta}
\alias{parse_imgt_fasta}
\title{Extract gene names from a reference FASTA.}
\usage{
parse_imgt_fasta(infile)
}
\arguments{
\item{infile}{A string, the path to FASTA file.}
}
\value{
A character vector of gene names.
}
\description{
Extract gene names from a reference FASTA.
}
