% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lookup.R
\name{parse_mixcr_csv}
\alias{parse_mixcr_csv}
\title{Extact gene names from a reference CSV for a given species}
\usage{
parse_mixcr_csv(species)
}
\arguments{
\item{species}{A string specifying the species, one of "human", "mouse",
or "rhesus".}
}
\value{
A character vector of gene names.
}
\description{
\code{parse_mixcr_csv()} generates mixcr gene name data for a given species.
}
\examples{
parse_mixcr_csv("human")
}
\keyword{internal}
