% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{which_frm_cols}
\alias{which_frm_cols}
\title{Determine input columns to use for converting gene names.}
\usage{
which_frm_cols(df, frm, frm_cols = NULL, verbose)
}
\arguments{
\item{df}{Dataframe containing TCR gene names.}

\item{frm}{A string, the input format of TCR data. Must be one of "tenx", "adaptive", "adaptivev2", or "imgt".}

\item{frm_cols}{A character vector, the custom column names to use.}

\item{verbose}{A boolean, whether to show messages.}
}
\value{
A character vector, column names to use.
}
\description{
Determine input columns to use for converting gene names.
}
\examples{
tcr_file <- get_example_path('tenx.csv')
df <- read.csv(tcr_file)
TCRconvertR:::which_frm_cols(df, 'tenx')
}
