% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lookup.R
\name{add_dash_one}
\alias{add_dash_one}
\title{Add a \code{-01} to genes without IMGT gene-level designation.}
\usage{
add_dash_one(gene_str)
}
\arguments{
\item{gene_str}{A string, the gene name.}
}
\value{
A string, the updated gene name.
}
\description{
Add a \code{-01} to genes without IMGT gene-level designation.
}
