% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{choose_lookup}
\alias{choose_lookup}
\title{Determine which lookup table to use and get filepath.}
\usage{
choose_lookup(frm, to, species = "human", verbose = TRUE)
}
\arguments{
\item{frm}{A string, the input format of TCR data. Must be one of "tenx", "adaptive", "adaptivev2", or "imgt".}

\item{to}{A string, the output format of TCR data. Must be one of "tenx", "adaptive", "adaptivev2", or "imgt".}

\item{species}{A string, the species. Optional; defaults to "human".}

\item{verbose}{A boolean, whether to show messages. Optional; defaults to TRUE}
}
\value{
A string, the path to correct lookup table.
}
\description{
Determine which lookup table to use and get filepath.
}
\examples{
choose_lookup("imgt", "adaptive")
}
\keyword{internal}
