% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_gene}
\alias{convert_gene}
\title{Convert T-cell receptor V, D, J, and/or C gene names from one naming convention to another.}
\usage{
convert_gene(df, frm, to, species = "human", frm_cols = NULL, verbose = TRUE)
}
\arguments{
\item{df}{Dataframe containing TCR gene names.}

\item{frm}{A string, the input format of TCR data. Must be one of "tenx", "adaptive", "adaptivev2", or "imgt".}

\item{to}{A string, the output format of TCR data. Must be one of "tenx", "adaptive", "adaptivev2", or "imgt".}

\item{species}{A string, the species folder name under \verb{tcrconvertr/inst/extdata/}. Optional; defaults to "human".}

\item{frm_cols}{A character vector of custom V/D/J/C gene column names. Optional; defaults to NULL.}

\item{verbose}{A boolean, whether to show messages. Optional; defaults to TRUE}
}
\value{
A dataframe of converted TCR data.
}
\description{
Convert T-cell receptor V, D, J, and/or C gene names from one naming convention to another.
}
\examples{
tcr_file <- get_example_path('tenx.csv')
df <- read.csv(tcr_file)
convert_gene(df, 'tenx', 'adaptive', verbose = FALSE)
}
