% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_example_path}
\alias{get_example_path}
\title{Get full path to given example file or directory.}
\usage{
get_example_path(file_name)
}
\arguments{
\item{file_name}{A string, the name of the example file or directory.}
}
\value{
A string, the path to example file or directory.
}
\description{
Get full path to given example file or directory.
}
