% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lookup.R
\name{pad_single_digit}
\alias{pad_single_digit}
\title{Add a zero to single-digit gene-level designatinon in gene names.}
\usage{
pad_single_digit(gene_str)
}
\arguments{
\item{gene_str}{A string, the gene name.}
}
\value{
A string, the updated gene name.
}
\description{
Add a zero to single-digit gene-level designatinon in gene names.
}
