% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lookup.R
\name{parse_mixcr_csv}
\alias{parse_mixcr_csv}
\title{Extract gene names from a reference CSV for a given species}
\usage{
parse_mixcr_csv(data_dir)
}
\arguments{
\item{data_dir}{A string, the path to directory containing mixcr CSV files.}
}
\value{
A character vector of gene names.
}
\description{
\code{parse_mixcr_csv()} generates mixcr gene name data for a given species.
}
\examples{
mixcr_dir <- get_example_path("mixcr_dir/test_mixcr")
parse_mixcr_csv(mixcr_dir)
}
\keyword{internal}
